script combines individual player csvs into 1 csv.
The script aligning_times.R takes a large set of individual player CSV
files and combines them together into a single master CSV file. Most
importantly, aligning_times.R aligns the times of the individual
player data, so that each row of the resulting CSV corresponds to a
single second in time, with each player's data represented on that row.